+++ Code/Runtime/Listeners/Non-Mono/IGameLoadListener.cs
@@ -14,11 +14,14 @@
  * If not, see <https://www.gnu.org/licenses/>. 
  */
 
+using UnityEngine.Scripting;
+
 namespace CarterGames.Assets.SaveManager
 {
     /// <summary>
     /// Defines an load listener implementation.
     /// </summary>
+    [RequireImplementors]
     public interface IGameLoadListener
     {
         /* —————————————————————————————————————————————————————————————————————————————————————————————————————————————